+++ src/Container/BasicContainer.php
@@ -36,7 +36,7 @@ class BasicContainer implements ContainerInterface
      *
      * @param string $id Identifier of the entry to look for.
      *
-     * @return boolean
+     * @return bool
      */
     public function has($id)
     {