+++ src/Dispatcher/ClassDispatcher.php
@@ -2,7 +2,6 @@
 namespace WoohooLabs\Harmony\Dispatcher;
 
 use Interop\Container\ContainerInterface;
-use WoohooLabs\Harmony\Config;
 use WoohooLabs\Harmony\Request\RequestInterface;
 use WoohooLabs\Harmony\Response\ResponseInterface;
 
@@ -12,11 +11,6 @@ class ClassDispatcher extends AbstractDispatcher
      * @var \Interop\Container\ContainerInterface
      */
     private $container;
-
-    /**
-     * @var \WoohooLabs\Harmony\Config
-     */
-    private $config;
 
     /**
      * @var string
@@ -29,13 +23,14 @@ class ClassDispatcher extends AbstractDispatcher
     private $methodName;
 
     /**
+     * @param \Interop\Container\ContainerInterface $container
      * @param string $className
      * @param string $methodName
-     * @param array $parameters
+     * @param array $params
      */
-    public function __construct($className, $methodName, array $parameters)
+    public function __construct(ContainerInterface $container, $className, $methodName, array $params)
     {
-        parent::__construct($parameters);
+        parent::__construct($params);
         $this->className = $className;
         $this->methodName= $methodName;
     }
@@ -49,17 +44,25 @@ class ClassDispatcher extends AbstractDispatcher
     {
         $object= $this->container->get($this->className);
 
-        if($this->config->getPreHandlerHookName() !== null && method_exists($object, $this->config->getPreHandlerHookName()) === true) {
-            $object->{$this->config->getPreHandlerHookName($request, $response)};
+        if(method_exists($object, "preHook") === true) {
+            $object->preHook($request, $response);
         }
 
         $response= $object->{$this->methodName}($request, $response);
 
-        if($this->config->getPostHandlerHookName() !== null && method_exists($object, $this->config->getPostHandlerHookName()) === true) {
-            $object->{$this->config->getPostHandlerHookName($request, $response)};
+        if(method_exists($object, "postHook") === true) {
+            $object->postHook($request, $response);
         }
 
         return $response;
+    }
+
+    /**
+     * @param \Interop\Container\ContainerInterface $container
+     */
+    public function setContainer(ContainerInterface $container)
+    {
+        $this->container= $container;
     }
 
     /**
@@ -71,6 +74,14 @@ class ClassDispatcher extends AbstractDispatcher
     }
 
     /**
+     * @param string $className
+     */
+    public function setClassName($className)
+    {
+        $this->className = $className;
+    }
+
+    /**
      * @return string
      */
     public function getMethodName()
@@ -79,18 +90,10 @@ class ClassDispatcher extends AbstractDispatcher
     }
 
     /**
-     * @param \WoohooLabs\Harmony\Config $config
+     * @param string $methodName
      */
-    public function setConfig(Config $config)
+    public function setMethodName($methodName)
     {
-        $this->config= $config;
-    }
-
-    /**
-     * @param \Interop\Container\ContainerInterface $container
-     */
-    public function setContainer(ContainerInterface $container)
-    {
-        $this->container= $container;
+        $this->methodName = $methodName;
     }
 }