+++ src/Router/MethodNotAllowedException.php
@@ -3,5 +3,4 @@ namespace WoohooLabs\Harmony\Router;
 
 class MethodNotAllowedException extends \Exception
 {
-
 }