+++ src/Router/RouteNotFoundException.php
@@ -3,5 +3,4 @@ namespace WoohooLabs\Harmony\Router;
 
 class RouteNotFoundException extends \Exception
 {
-
 }