+++ src/Serializer/DeserializerInterface.php
@@ -14,7 +14,7 @@ interface DeserializerInterface
 
     /**
      * @param string $format
-     * @return boolean
+     * @return bool
      */
     public function isFormatDeserializable($format);
 }