+++ src/Serializer/SerializerInterface.php
@@ -14,7 +14,7 @@ interface SerializerInterface
 
     /**
      * @param string $format
-     * @return boolean
+     * @return bool
      */
     public function isFormatSerializable($format);
 }